Update user_added_playlists_page.dart
Modified so Custom Playlists load again and custom playlists are removable! 
The problem was that the conversion of "yitd" to String made it so that it's content is 'null' instad of null, which rendered multiple checks useless.
My guess is that the change was done because the playlistRemoveFunction requiers a non nullable string. Therefore a new (working) remove function was also written.